Validate edit input to prevent duplicate and reorganize file structure
